LumosX/requirements.txt
absl-py==2.1.0
accelerate==1.3.0
addict==2.4.0
aiofiles==23.2.1
aiohappyeyeballs==2.4.6
aiohttp==3.11.12
aiosignal==1.3.2
aliyun-python-sdk-core==2.16.0
aliyun-python-sdk-kms==2.16.5
annotated-types==0.7.0
anyio==4.8.0
asttokens==3.0.0
async-timeout==5.0.1
attrs==25.1.0
av==14.2.0
backcall==0.2.0
backoff==2.2.1
beautifulsoup4==4.13.3
bleach==6.3.0
bs4==0.0.2
certifi==2025.1.31
cffi==1.17.1
cfgv==3.4.0
charset-normalizer==3.4.1
clean-fid==0.1.35
click==8.1.8
contourpy==1.3.1
crcmod==1.7
cryptography==44.0.2
cycler==0.12.1
dashscope==1.22.1
debugpy==1.8.14
decorator==5.2.1
decord==0.6.0
deepspeed==0.16.3
defusedxml==0.7.1
diffusers==0.32.1
dill==0.3.9
distlib==0.3.9
distro==1.9.0
docker-pycreds==0.4.0
docopt==0.6.2
easydict==1.13
einops==0.8.0
exceptiongroup==1.2.2
executing==2.2.0
fastapi==0.115.11
fastjsonschema==2.21.2
ffmpy==0.5.0
filelock==3.13.1
fonttools==4.56.0
frozenlist==1.5.0
fsspec==2024.6.1
ftfy==6.3.1
gitdb==4.0.12
GitPython==3.1.44
gradio==5.20.0
gradio_client==1.7.2
groovy==0.1.2
grpcio==1.70.0
h11==0.14.0
hjson==3.1.0
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.28.1
identify==2.6.8
idna==3.10
imageio==2.36.1
imageio-ffmpeg==0.6.0
importlib_metadata==8.6.1
inquirerpy==0.3.4
ipdb==0.13.13
ipython==8.12.3
jedi==0.19.2
Jinja2==3.1.4
jiter==0.8.2
jmespath==0.10.0
joblib==1.4.2
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
jupyter_client==8.8.0
jupyter_core==5.9.1
jupyterlab_pygments==0.3.0
kiwisolver==1.4.8
loguru==0.7.3
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.10.1
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.2.0
mmengine==0.10.7
mpmath==1.3.0
msgpack==1.1.0
multidict==6.1.0
nbclient==0.10.4
nbconvert==7.16.6
nbformat==5.10.4
networkx==3.3
ninja==1.11.1.3
nodeenv==1.9.1
numpy==1.24.4
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-ml-py==12.570.86
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
openai==1.65.2
openai-clip==1.0.1
opencv-python==4.11.0.86
orjson==3.10.15
oss2==2.19.1
packaging==24.2
pandarallel==1.6.5
pandas==2.2.3
pandocfilters==1.5.1
parso==0.8.4
peft==0.15.0
pexpect==4.9.0
pfzy==0.3.4
pickleshare==0.7.5
pillow==11.0.0
pipreqs==0.5.0
platformdirs==4.3.6
pre_commit==4.1.0
proglog==0.1.10
prompt_toolkit==3.0.50
propcache==0.2.1
protobuf==5.29.3
psutil==6.1.1
ptyprocess==0.7.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
pyarrow==19.0.1
pycparser==2.22
pycryptodome==3.21.0
pydantic==2.10.6
pydantic_core==2.27.2
pydub==0.25.1
Pygments==2.19.1
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-multipart==0.0.20
pytz==2025.1
PyYAML==6.0.2
pyzmq==27.1.0
qwen-vl-utils==0.0.14
referencing==0.37.0
regex==2024.11.6
requests==2.32.3
rich==13.9.4
rotary-embedding-torch==0.8.6
rpdb==0.2.0
rpds-py==0.30.0
ruff==0.9.9
safehttpx==0.1.6
safetensors==0.5.2
scikit-learn==1.6.1
scipy==1.15.2
seaborn==0.13.2
semantic-version==2.10.0
sentencepiece==0.2.0
sentry-sdk==2.22.0
setproctitle==1.3.5
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.6
stack-data==0.6.3
starlette==0.46.0
sympy==1.13.1
tensorboard==2.19.0
tensorboard-data-server==0.7.2
termcolor==2.5.0
threadpoolctl==3.6.0
tiktoken==0.9.0
timm==1.0.15
tinycss2==1.4.0
tokenizers==0.21.0
tomli==2.2.1
tomlkit==0.13.2
torch==2.6.0+cu124
torchaudio==2.6.0
torchvision==0.21.0+cu124
tornado==6.5.4
tqdm==4.67.1
traitlets==5.14.3
transformers==4.48.3
triton==3.2.0
typer==0.15.2
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
uvicorn==0.34.0
virtualenv==20.29.2
wandb==0.20.1
wcwidth==0.2.13
webencodings==0.5.1
websocket-client==1.8.0
websockets==15.0
Werkzeug==3.1.3
xformers==0.0.29.post3
yapf==0.43.0
yarg==0.1.9
yarl==1.18.3
zipp==3.21.0